Ya hice los ejercicios que se requerian en Cisharp

+++ ejercicioEvaluacion/ejercicioExamen/clases_y_atributos/ICalculable.cs
@@ -9,6 +9,8 @@ namespace ejercicioExamen.clases_y_atributos
     public interface ICalculable
     {
         //Definimos un método para implementarlo en Empleado
-        int CalcularSalario();
+        decimal CalcularSalario();
+
+        
     }
 }
